requirements.txt
flask>=3.0.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
pyzbar>=0.1.9
zxing-cpp>=2.2.0
python-dateutil>=2.9.0
werkzeug>=3.0.0
numpy>=1.26.0
